+++ packages/babel-plugin-veui/lib/index.js
@@ -4,6 +4,8 @@ Object.defineProperty(exports, "__esModule", {
   value: true
 });
 
+var _slicedToArray = function () { function sliceIterator(arr, i) { var _arr = []; var _n = true; var _d = false; var _e = undefined; try { for (var _i = arr[Symbol.iterator](), _s; !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i["return"]) _i["return"](); } finally { if (_d) throw _e; } } return _arr; } return function (arr, i) { if (Array.isArray(arr)) { return arr; } else if (Symbol.iterator in Object(arr)) { return sliceIterator(arr, i); } else { throw new TypeError("Invalid attempt to destructure non-iterable instance"); } }; }();
+
 exports.default = function (babel) {
   var t = babel.types;
 
@@ -30,7 +32,12 @@ exports.default = function (babel) {
               path.insertBefore(t.importDeclaration([t.importDefaultSpecifier(t.identifier(local.name))], t.stringLiteral(src)));
               path.getSibling(path.key - 1).stop();
             } else {
-              var componentSrc = getComponentPath(imported.name);
+              var realName = getComponentName(imported.name);
+
+              var componentSrc = getComponentPath(realName);
+              if (!componentSrc) {
+                throw new Error('[' + realName + '] is not a valid component in VEUI.');
+              }
               var name = local.name || imported.name;
               path.insertBefore(t.importDeclaration([t.importDefaultSpecifier(t.identifier(name))], t.stringLiteral(componentSrc)));
               path.getSibling(path.key - 1).stop();
@@ -53,6 +60,16 @@ var _components2 = _interopRequireDefault(_components);
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 var COMPONENTS_DIRNAME = 'components';
+
+var VAR_PATTERN = /(?:V(?:eui)?)?([A-Z][a-zA-Z]*)/;
+
+function getComponentName(importedName) {
+  var _importedName$match = importedName.match(VAR_PATTERN),
+      _importedName$match2 = _slicedToArray(_importedName$match, 2),
+      name = _importedName$match2[1];
+
+  return name || null;
+}
 
 function getComponentPath(componentName) {
   var entry = _components2.default.find(function (_ref2) {